requirements.txt
# Core Framework
streamlit==1.29.0
gradio==4.8.0

# Multi-Agent Framework
crewai==0.11.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
openai==1.7.2


# Vision AI
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
ultralytics==8.0.226
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git
torch==2.1.1
torchvision==0.16.1
Pillow==10.1.0

# Geospatial
geopandas==0.14.1
rasterio==1.3.9
folium==0.15.1
earthengine-api==0.1.384
pyproj==3.6.1
shapely==2.0.2

# Web Scraping
selenium==4.16.0
beautifulsoup4==4.12.2
requests==2.31.0
webdriver-manager==4.0.1

# Optimization
scipy==1.11.4
numpy==1.26.2
scikit-learn==1.3.2
deap==1.4.1

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0
kaleido==0.2.1

# Data Processing
pandas==2.1.4
openpyxl==3.1.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
tqdm==4.66.1
geopy==2.4.1
timezonefinder==6.2.0
